move update functors to separate folder
