+++ hydra/include/hydra/backend/updates/update_block_traversability_functor.h
