+++ hydra/include/hydra/backend/updates/update_buildings_functor.h
