+++ hydra/include/hydra/backend/updates/update_region_growing_traversability_functor.h
