+++ hydra/src/backend/CMakeLists.txt
@@ -6,20 +6,21 @@ target_sources(
           ${CMAKE_CURRENT_SOURCE_DIR}/deformation_interpolator.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/dsg_updater.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/external_loop_closure_receiver.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/generic_update_functor.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/merge_proposer.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/merge_tracker.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/mesh_clustering.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/mst_factors.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_agents_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_block_traversability_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_buildings_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_frontiers_functor.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/update_functions.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_objects_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_places_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_region_growing_traversability_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_rooms_functor.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/update_surface_places_functor.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/zmq_interfaces.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/generic_update_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/mesh_clustering.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_agents_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_block_traversability_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_buildings_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_frontiers_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_objects_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_places_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_region_growing_traversability_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_rooms_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/update_surface_places_functor.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/updates/zmq_room_label_updater.cpp
 )
