+++ hydra/src/backend/updates/update_region_growing_traversability_functor.cpp
@@ -32,17 +32,13 @@
  * Government is authorized to reproduce and distribute reprints for Government
  * purposes notwithstanding any copyright notation herein.
  * -------------------------------------------------------------------------- */
-#include "hydra/backend/update_region_growing_traversability_functor.h"
+#include "hydra/backend/updates/update_region_growing_traversability_functor.h"
 
 #include <config_utilities/config.h>
 #include <config_utilities/validation.h>
 #include <spark_dsg/node_attributes.h>
 #include <spark_dsg/traversability_boundary.h>
 
-#include <queue>
-
-#include "hydra/common/global_info.h"
-#include "hydra/utils/nearest_neighbor_utilities.h"
 #include "hydra/utils/timing_utilities.h"
 
 namespace hydra {
